Fixed dependency issues with modules.
[Done] Moved runtime constants.
All should be ready for move from users and projects.
