+++ pallets/chocolate/src/benchmarking.rs
@@ -6,15 +6,15 @@ use super::*;
 use crate::Pallet as Chocolate;
 use frame_benchmarking::{benchmarks, impl_benchmark_test_suite, whitelisted_caller};
 use frame_system::RawOrigin;
-
-benchmarks! {
-	do_something {
-		let s in 0 .. 100;
-		let caller: T::AccountId = whitelisted_caller();
-	}: _(RawOrigin::Signed(caller), s)
-	verify {
-		assert_eq!(Something::<T>::get(), Some(s));
-	}
+// calls do_something multiple times 0..100 inputs verifying it got stored each time
+benchmarks! { // comment for now, do_something is gone, but it's legacy remains
+	// do_something {
+		// let s in 0 .. 100;
+		// let caller: T::AccountId = whitelisted_caller();
+	// }: _(RawOrigin::Signed(caller), s)
+	// verify {
+		// assert_eq!(Something::<T>::get(), Some(s));
+	// }
 }
 
 impl_benchmark_test_suite!(Chocolate, crate::mock::new_test_ext(), crate::mock::Test);
