+++ pallets/users/Cargo.toml
@@ -64,6 +64,11 @@ default-features = false
 git = 'https://github.com/paritytech/substrate.git'
 tag = 'monthly-2021-08'
 version = '4.0.0-dev'
+# added dep
+[dependencies.chocolate-users]
+path = '../../primitives/chocolate-users'
+default-features = false
+version = '0.1.0'
 
 [features]
 default = ['std']
@@ -74,5 +79,7 @@ std = [
     'frame-support/std',
     'frame-system/std',
     'frame-benchmarking/std',
+     # added
+    'chocolate-users/std',
 ]
 try-runtime = ['frame-support/try-runtime']
