--- primitives/chocolate-node-constants/src/tests.rs
@@ -1,5 +1,0 @@
-#[test]
-fn it_works() {
-	let result = 2 + 2;
-	assert_eq!(result, 4);
-}
