+++ primitives/chocolate-projects/src/lib.rs
@@ -1,8 +1,3 @@
-#[cfg(test)]
-mod tests {
-	#[test]
-	fn it_works() {
-		let result = 2 + 2;
-		assert_eq!(result, 4);
-	}
-}
+#![cfg_attr(not(feature = "std"), no_std)]
+
+
