+++ runtime/src/lib.rs
@@ -54,8 +54,7 @@ pub use sp_runtime::{Perbill, Percent, Permill};
 pub use pallet_chocolate;
 /// Import the users pallet
 pub use pallet_users;
-/// An index to a block.
-pub type BlockNumber = u32;
+pub use chocolate_node_constants::*;
 
 /// Alias to 512-bit hash when used in the context of a transaction signature on the chain.
 pub type Signature = MultiSignature;
@@ -63,9 +62,6 @@ pub type Signature = MultiSignature;
 /// Some way of identifying an account on the chain. We intentionally make it equivalent
 /// to the public key of our transaction signing scheme.
 pub type AccountId = <<Signature as Verify>::Signer as IdentifyAccount>::AccountId;
-
-/// Balance of an account.
-pub type Balance = u128;
 
 /// Index of a transaction in the chain.
 pub type Index = u32;
@@ -117,27 +113,7 @@ pub const VERSION: RuntimeVersion = RuntimeVersion {
 
 // Note: This could be later moved into a constants file - Separating time and currency, and importing primitives from node-primitives. Or here.
 
-/// This determines the average expected block time that we are targeting.
-/// Blocks will be produced at a minimum duration defined by `SLOT_DURATION`.
-/// `SLOT_DURATION` is picked up by `pallet_timestamp` which is in turn picked
-/// up by `pallet_aura` to implement `fn slot_duration()`.
-///
-/// Change this to adjust the block time.
-pub const MILLISECS_PER_BLOCK: u64 = 6000;
 
-// NOTE: Currently it is not possible to change the slot duration after the chain has started.
-//       Attempting to do so will brick block production.
-pub const SLOT_DURATION: u64 = MILLISECS_PER_BLOCK;
-
-// Time is measured by number of blocks.
-pub const MINUTES: BlockNumber = 60_000 / (MILLISECS_PER_BLOCK as BlockNumber);
-pub const HOURS: BlockNumber = MINUTES * 60;
-pub const DAYS: BlockNumber = HOURS * 24;
-
-// This specifies currency things. Currently set to recurse on our currency. CHANGE FOR DEFI
-pub const MILLICENTICHOC: Balance = 1_000_000_000;
-pub const CENTICHOC: Balance = 1_000 * MILLICENTICHOC;
-pub const CHOC: Balance = 100 * CENTICHOC;
 pub const fn deposit(items: u32, bytes: u32) -> Balance {
 	items as Balance * 15 * CENTICHOC + (bytes as Balance) * 6 * CENTICHOC
 }
